Adjust make-changelog to retrieve the pull-requests from pagure correctly
Currently make-changelog could not retrieve the pull-requests from pagure
since there are most often no merge commit.
To circumvent this situation, we get the date of the first commit in the
series and get all the pull-request that have been merged since then.
As the updated_on field might be missing, we drop back to date_created
in these cases.
This also gives the opportunity to adjust the script more broadly for
github vs pagure.